Replaced floats with doubles (#355)
This makes our APIs more consistent. With optimizations enabled, doubles are just as efficient as floats on ARMv7, so we should take advantage of the extra precision.

+++ wpilibc/athena/include/ADXRS450_Gyro.h
@@ -33,7 +33,7 @@ class ADXRS450_Gyro : public GyroBase {
   explicit ADXRS450_Gyro(SPI::Port port);
   virtual ~ADXRS450_Gyro() = default;
 
-  float GetAngle() const override;
+  double GetAngle() const override;
   double GetRate() const override;
   void Reset() override;
   void Calibrate() override;
