+++ wpilibc/athena/src/MotorSafetyHelper.cpp
@@ -59,7 +59,7 @@ void MotorSafetyHelper::Feed() {
  * Set the expiration time for the corresponding motor safety object.
  * @param expirationTime The timeout value in seconds.
  */
-void MotorSafetyHelper::SetExpiration(float expirationTime) {
+void MotorSafetyHelper::SetExpiration(double expirationTime) {
   std::lock_guard<priority_recursive_mutex> sync(m_syncMutex);
   m_expiration = expirationTime;
 }
@@ -68,7 +68,7 @@ void MotorSafetyHelper::SetExpiration(float expirationTime) {
  * Retrieve the timeout value for the corresponding motor safety object.
  * @return the timeout value in seconds.
  */
-float MotorSafetyHelper::GetExpiration() const {
+double MotorSafetyHelper::GetExpiration() const {
   std::lock_guard<priority_recursive_mutex> sync(m_syncMutex);
   return m_expiration;
 }
